requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
websockets==12.0
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
aiosqlite==0.20.0
redis==5.0.3
httpx==0.27.0